Chaoscrawl
A version of Crawl with a permanent Misdirect effect. All monsters appear as some
other type - e.g. all rats might be goblins in one game, or hydras in another, or
The Royal Jelly in a third. (Exceptions are tentacles & player ghosts.) Can you
figure out which is which? Maybe!