Allow phasing out support for old minor versions
Crawl currently supports saves from over 200 minor versions. Rather than
breaking save compatibility entirely with a TAG_MAJOR_VERSION bump, a
less disruptive approach is to incrementally remove support for older
minor versions, giving a 'rolling compatibility' model. With such a
model, the latest crawl binary might maintain support for ~3 years of
minor versions; the period of support is, of course, flexible.

This commit switches to using the TAG_MINOR_INVALID enum value to
determine the minimum supported minor version. This means that this
commit breaks compatibility with 0.11 (save format 33.17).

When phasing out support for old saves, the intended workflow is that n
tags are removed from the start of the enum list, and TAG_MINOR_INVALID
is incremented by n, so that all remaining tags keep the same value.
There is a unit test to help ensure that no accidental drift occurs.

One complication is that server game transfer scripts assume that
compatibility is determined purely by the major version, and so a
too-old save file may be transferred to a binary which will refuse to
load it, requiring manual invervention by an administrator. This should
not happen very often, and various strategies exist, e.g. when removing
support for a version, force upgrade associated saves to the next-oldest
binary, instead of the newest.