Re-implement boulder beetles
Boulder beetles return as a speed 10, 12 HD monster with good AC, melee
slightly better than a death yak, and a rolling attack that doubles
their movement speed and melee damage (1d32->1d64).

They will only self-enchant with the rolling debuff when not already
adjacent to their foe, hence either killing them at range or getting
them into melee range is a priority to avoid their 1d64 rolling damage.
After performing a melee attack in the rolling state, the beetle always
debuffs, returning to its unrolled state. One of the advantages of
having the roll attack be normal melee is that it uses normal melee
mechanics, checking SH and EV, and their modified attack damage is shown
under x-v when in the rolling state.

Boulder beetles also have a blink ability in order to eventually reset
their positioning and allow a subsequent roll, although they cast this
with low frequency. While rolling, they never use either of their
abilities.

This is a WIP implementation with a few outstanding issues.

TODO:

1) "Missing" the player.  In my original proposal, I wanted a missed
roll to allow them to "roll past" the player, giving the player a chance
for additional punishment but also allowing the beetle to set up another
roll. Currently this is not implemented, since this aspect was mostly to
the old trivially avoidable (but comedic) roll attack. They already can
reset their position using blink, and with the profile of this monster,
giving the player additional opportunities to damage it would likely
just make it too easy. Still, it remains an option if `??meme rule` is
being ignored.

2) Cool-down after performing a roll? Their being unable to roll when
adjacent is a kind of natural cool-down. I'm not sure any extra
cool-down is necessary, as they're not too hard of a monster as things
are.

3) Rolling duration. Right now it's 10 to 15 turns, giving faster
players a fighting small chance at outrunning a rolling beetle that
starts rolling many tiles away, but for normal speed players, even being
7 tiles away means they will not outrun the roll.

4) AI changes while rolling. Should beetles ignore damaging terrain like
traps and clouds while rolling? Currently they don't change their damage
consideration compared to their normal state.

5) Theme of "blink". This is done as the usual translocations effect,
although I had originally planned for this to be themed as a "flip" to
make more thematic sense as a natural ability. Jumping spiders likewise
have no special translocations theme and do the same kind of blink, so
maybe leaving the normal blink is best.

6) Placement and vaults. I mostly re-added them to the spawn lists with
sensible weights and to vaults and Sprint in the appropriate places. I
need to make sure these additions are correct and see if they should be
added in other places. I'll likely make some special boulder beetle
transporter or runed door vaults that get at the theme of the old
spoilery boulder beetle trap vaults, just without the spoilers and
trivial damage avoidance.

7) Add description for Roll ability and rewrite description of monster
itself.