New Okawaru ability: Duel
Granted at 5* (with Finesse moving earlier, to 4*), the Duel ability
smite-targets a single non-summoned enemy in LOS (which must be threat
level "tough" or higher), and transports them and the player to an Arena
portal vault. Water/lava-only monsters cannot be dueled, and banishment
from within the Arena is blocked.

Winning the duel opens a gate to return from the portal vault, and the
player will be forcibly teleported back after a short duration (to give
time to loot).

TODO: fancy up the Arena vault (and add variations?), add more spectator
speech lines, give spectators a better name and tiles, add ability icon.