Add some Arena vault variations
These should remain simple, with minor cosmetic terrain at most, and should
be small enough for the whole arena to be in LOS on arrival. Also move
some vault setup into Lua to reduce duplication, and prevent destruction of
the Arena's metal walls (it's amusing to be able to break out and kill the
spectators with Shatter, but it also restricts vault design to require all
the terrain surrounding the exit gate to be permaglass).